Restart version control workflow for project
Add the project code in order to manage a more granular version-
control workflow. At this point, the project has been entirely re-
organized. It is no longer orchestrated by a series of notebooks, but
rather a dedicated 'src' directory that contains extractions, pre-
processing scripts, raw, clean, merged, and intermediate data, as well
as a CLI to run the project from a global or modular stance. The next
step here is to build an extraction script for nomi-sec's PoC-in-GitHub
database, to use the dates retrieved therefrom to capture EPSS data
gathered from FIRST, and to merge all this data into one table in order
to conduct some preliminary analyses about the correlation between
exploited CVEs as marked by CISA's KEV catalogue and the number of
proof-of-concept exploits.