Update CVE ID validation function
Add logic to accept a fallback argument that may contain CVE IDs if
the primary location does not. Ensure the function checks for IDs
that are not necessarily the beginning and end of the strings passed
to the function, makes sure that 'CVE' is uppercase, and makes sure
that the ordinal group within each ID is zero-padded to at least four
digits.

A minimal amount of work was also put into the proof-of-concept
cleaning/preprocessing script.